Refactor repo into a non-virtual cargo workspace with a root package.
The old way of using a virtual workspace is incorrect for an application like this.
It also prevents existing bundle tools (cargo-bundle) from working with Moxin because they all expect a root package to be defined.

Technically we don't even need a workspace at all for this repo, but I left it in just for the sake of posterity.