Make Moxin runnable from a macOS `.app` bundle.
The problem is that when an app bundle is run, the current working directory is `/` (root), which is read-only on macOS.
Thus, the attempt to access `./data.sqlite` is actually an attempt to create the file `/data.sqlite`, which immediately fails
because a file cannot be created on the read-only root partition.

Instead of using the current working directory, this commit now changes file paths to be relative to `$HOME`,
which will only really work properly on Unix-like environments where the `$HOME` environment variable is defined.
Even on Unix-like systems, this isn't *guaranteed* to work, since various sandboxing or environment masking techniques may be in effect.

Note: do not merge this yet, it's messy.