build: add flatpak packaging
Add a maintenance Makefile that can build Linux Flatpaks for Moxin. This
allows running Moxin isolated from the Linux host, using the standard
Flatpak sandboxing technology.

The moxin build runs outside of the Flatpak sandbox for now. This can
lead to ABI issues, if the host build is not compatible with the
freedesktop SDK used inside of the Flatpak. However, building inside of
the Flatpak sandbox requires vendoring of all dependencies, which is a
hassle to set up. This can be adjusted in the future.

Signed-off-by: David Rheinsberg <david@readahead.eu>