Do not refresh all downloaded info from backend upon cancel/pause download
The backend implementation spends some time before adding a new download
to the DB table when pending downloads are tracked, so it is not safe to
refresh the whole list of pending downloads after cancelling/pausing.
So, we can simply update manually the frontend copy of the information
rather than polling it from the backend.