Do not show cancel/pause buttons when downloads are being initialized on backend
This is needed due to a current limitation in the backend
implementation. If frontend sends a PauseDownload or CancelDownload
command right after a DownloadFile, it doesn't have any effect. It
takes some time (sometimes a few seconds) to be "listening" for those
commands, so the frontend needs to adapt to it until we have better
solutions.