Fix intermittent problem in ModelCard on title or description
There was an issue where the title or description of some models in the
main listing were not visible. The root problem was we were using
similar identifiers for the title and description labels in ModelCard
and ModelCardViewAllModal widgets (which is a nested view under the
model card view).

The solution is to use unique identifiers, so the widget lookup
retrieves what we expect in each case..