moxin-runner: explicitly detect CUDA on Linux
The WasmEdge `install_v2.sh` script currently treats the `--noavx`
option as a hard override that bypasses CUDA detection.
Thus, we must manually check for CUDA first, and if CUDA is detected,
we must *not* pass in the `--noavx` option.

We now only pass in the `--noavx` option if CUDA is not found
*and* if the CPU doesn't support AVX512.

It is not clear what WasmEdge will do if CUDA is detected but the
runtime determines that the graphics card is insufficient for its needs
-- will it revert back to CPU instructions that require AVX512??