MolyKit: Re-usable AI widget kit (#368)
Introduces a mew MolyKit crate with a more generic chat widget that can be re-used in other Makepad applications.
This is work in progress, and will be used for replacing the chat panel and http clients in Moly.

It also introduces moly-mini, a playground for MolyKit that will be either removed later or reworked into examples inside the moly-mini directory.

---------

Co-authored-by: Julian Montes de Oca <julian@wyeworks.com>